+++ src/spyglass/common/populate_all_common.py
@@ -1,59 +1,63 @@
-from spyglass.spikesorting.imported import ImportedSpikeSorting
-from spyglass.utils.dj_mixin import SpyglassMixin
-
-from .common_behav import (
+from spyglass.common.common_behav import (
     PositionSource,
     RawPosition,
     StateScriptFile,
     VideoFile,
 )
-from .common_dio import DIOEvents
-from .common_ephys import Electrode, ElectrodeGroup, Raw, SampleCount
-from .common_nwbfile import Nwbfile
-from .common_session import Session
-from .common_task import TaskEpoch
+from spyglass.common.common_dio import DIOEvents
+from spyglass.common.common_ephys import (
+    Electrode,
+    ElectrodeGroup,
+    Raw,
+    SampleCount,
+)
+from spyglass.common.common_nwbfile import Nwbfile
+from spyglass.common.common_session import Session
+from spyglass.common.common_task import TaskEpoch
+from spyglass.spikesorting.imported import ImportedSpikeSorting
+from spyglass.utils import logger
 
 
 def populate_all_common(nwb_file_name):
     # Insert session one by one
     fp = [(Nwbfile & {"nwb_file_name": nwb_file_name}).proj()]
-    print("Populate Session...")
+    logger.info("Populate Session...")
     Session.populate(fp)
 
     # If we use Kachery for data sharing we can uncomment the following two lines. TBD
-    # print('Populate NwbfileKachery...')
+    # logger.info('Populate NwbfileKachery...')
     # NwbfileKachery.populate()
 
-    print("Populate ElectrodeGroup...")
+    logger.info("Populate ElectrodeGroup...")
     ElectrodeGroup.populate(fp)
 
-    print("Populate Electrode...")
+    logger.info("Populate Electrode...")
     Electrode.populate(fp)
 
-    print("Populate Raw...")
+    logger.info("Populate Raw...")
     Raw.populate(fp)
 
-    print("Populate SampleCount...")
+    logger.info("Populate SampleCount...")
     SampleCount.populate(fp)
 
-    print("Populate DIOEvents...")
+    logger.info("Populate DIOEvents...")
     DIOEvents.populate(fp)
 
     # sensor data (from analog ProcessingModule) is temporarily removed from NWBFile
     # to reduce file size while it is not being used. add it back in by commenting out
     # the removal code in spyglass/data_import/insert_sessions.py when ready
-    # print('Populate SensorData')
+    # logger.info('Populate SensorData')
     # SensorData.populate(fp)
 
-    print("Populate TaskEpochs")
+    logger.info("Populate TaskEpochs")
     TaskEpoch.populate(fp)
-    print("Populate StateScriptFile")
+    logger.info("Populate StateScriptFile")
     StateScriptFile.populate(fp)
-    print("Populate VideoFile")
+    logger.info("Populate VideoFile")
     VideoFile.populate(fp)
-    print("RawPosition...")
+    logger.info("RawPosition...")
     PositionSource.insert_from_nwbfile(nwb_file_name)
     RawPosition.populate(fp)
 
-    print("Populate ImportedSpikeSorting...")
+    logger.info("Populate ImportedSpikeSorting...")
     ImportedSpikeSorting.populate(fp)