+++ src/spyglass/spikesorting/spikesorting_artifact.py
@@ -8,16 +8,15 @@ import scipy.stats as stats
 import spikeinterface as si
 from spikeinterface.core.job_tools import ChunkRecordingExecutor, ensure_n_jobs
 
-from spyglass.utils.dj_mixin import SpyglassMixin
-
-from ..common.common_interval import (
+from spyglass.common.common_interval import (
     IntervalList,
     _union_concat,
     interval_from_inds,
     interval_set_difference_inds,
 )
-from ..utils.nwb_helper_fn import get_valid_intervals
-from .spikesorting_recording import SpikeSortingRecording
+from spyglass.spikesorting.spikesorting_recording import SpikeSortingRecording
+from spyglass.utils import SpyglassMixin, logger
+from spyglass.utils.nwb_helper_fn import get_valid_intervals
 
 schema = dj.schema("spikesorting_artifact")
 
@@ -204,7 +203,7 @@ def _get_artifact_times(
             [[valid_timestamps[0], valid_timestamps[-1]]]
         )
         artifact_times_empty = np.asarray([])
-        print(
+        logger.info(
             "Amplitude and zscore thresholds are both None, skipping artifact detection"
         )
         return recording_interval, artifact_times_empty
@@ -220,7 +219,7 @@ def _get_artifact_times(
 
     # detect frames that are above threshold in parallel
     n_jobs = ensure_n_jobs(recording, n_jobs=job_kwargs.get("n_jobs", 1))
-    print(f"using {n_jobs} jobs...")
+    logger.info(f"using {n_jobs} jobs...")
 
     func = _compute_artifact_chunk
     init_func = _init_artifact_worker
@@ -260,7 +259,7 @@ def _get_artifact_times(
             [[valid_timestamps[0], valid_timestamps[-1]]]
         )
         artifact_times_empty = np.asarray([])
-        print("No artifacts detected.")
+        logger.warn("No artifacts detected.")
         return recording_interval, artifact_times_empty
 
     # convert indices to intervals